// src/API/BiliBiliAPI/BiliBiliSearch/SearchRequestInterface.ts
export interface SearchRequest
{
    code: number;
    message: string;
    ttl: number;
    data?: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        page_size: number; // 每页条数
        num_results: number; // 总条数，最大为1000
        num_pages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // 不知道有什么用途
        cost_time: cost_time;
        exp_list: exp_list; // 不知道有什么用途
        egg_hit: number; // 不知道有什么用途
        pageinfo: {
            video: pageInfoObject; // 视频数
            bangumi: pageInfoObject; // 不知道有什么用途
            special: pageInfoObject; // 不知道有什么用途
            topic: pageInfoObject; // 话题数
            upuser: pageInfoObject; // 不知道有什么用途
            tv: pageInfoObject; // 不知道有什么用途
            movie: pageInfoObject; // 不知道有什么用途
            media_bangumi: pageInfoObject; // 番剧数
            media_ft: pageInfoObject; // 电影数
            related_search: pageInfoObject;
            user: pageInfoObject; // 不知道有什么用途
            activity: pageInfoObject; // 活动数
            operation_card: pageInfoObject; // 不知道有什么用途
            pgc: pageInfoObject; // 不知道有什么用途
            live: pageInfoObject; // 直播间数
            live_all: pageInfoObject; // 不知道有什么用途
            live_user: pageInfoObject; // 主播数
            live_master: pageInfoObject; // 不知道有什么用途
            article: pageInfoObject; // 专栏数
            live_room: pageInfoObject; // 直播数
            bili_user: pageInfoObject; // 用户数
        }; // 分类页数信息	
        top_tlist: {
            video: number; // 视频数
            bangumi: number; // 番剧数
            special: number; // 还不知道是什么
            topic: number; // 还不知道是什么
            upuser: number; // 还不知道是什么
            tv: number; // 还不知道是什么
            movie: number; // 还不知道是什么
            card: number; // 还不知道是什么
            media_bangumi: number; // 番剧数
            media_ft: number; // 电影数
            pgc: number; // 还不知道是什么
            live: number; // 直播间数
            user: number; // 用户数
            activity: number; // 还不知道是什么
            operation_card: number; // 还不知道是什么
            live_user: number; // 主播数
            article: number; // 专栏数
            live_room: number; // 直播数
            live_master: number; // 还不知道是什么
            bili_user: number; // 用户数
        }; // 分类结果数目信息
        show_column: number; // 不知道有什么用途
        show_module_list: string[]; // 返回结果类型列表
        app_display_option: {
            is_search_page_grayed: number; // 不知道有什么用途
        }; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途
        result: {
            result_type: string; // 返回结果类型
            data: media_ftAndmedia_bangumiResult[] & VideoResult[] & UserResult[] & web_gameResult[] & activityResult[]; // 事情是这样的，他可能是这些的其中一个，反正自求多福
        }[];
        is_search_page_grayed: number; // 不知道有什么用途
    };
}
// ...
export interface SearchRequestByTypeVideo
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: VideoResult[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypeMediaBangumiAndMediaFT
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: media_ftAndmedia_bangumiResult[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypeLive
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        pageinfo: {
            live_user: pageInfoObject;
            live_room: pageInfoObject;
        };
        result: liveResult;
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypeLiveRoom
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: live_roomResult[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypeLiveUser
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: live_userResult[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypeArticle
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: ArticleResult[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

export interface SearchRequestByTypePhoto
{
    code: number;
    message: string;
    ttl: number;
    data: {
        seid: string; // 搜索id
        page: number; // 页数，固定为1
        pagesize: number; // 每页条数
        numResults: number; // 总条数，最大为1000
        numPages: number; // 分页数
        suggest_keyword: string; // 不知道有什么用途
        rqt_type: string; // search
        cost_time: cost_time;
        exp_list: exp_list;
        egg_hit: number; // 不知道有什么用途
        result: {}[];
        show_column: number; // 不知道有什么用途
        in_black_key: number; // 不知道有什么用途
        in_white_key: number; // 不知道有什么用途

    };
}

interface pageInfoObject
{
    total: number; // 总计数量
    numResults: number; // 总计数量
    pages: number; // 分页数量
}

interface media_ftAndmedia_bangumiResult
{
    type: string; // 类型
    media_id: number; // 剧集mdid
    title: string; // 剧集标题
    org_title: string; // 剧集原名
    media_type: number; // 剧集类型 1：番剧 2：电影 3：纪录片 4：国创 5：电视剧 7：综艺
    cv: string; // 声优
    staff: string; // 制作人员
    season_id: number; // 剧集ssid
    is_avid: boolean; // 不知道有什么用途
    hit_columns: string[] | null; // 关键字匹配类型
    hit_epids: string; // 关键字匹配分集标题的分集epid
    season_type: number; // 剧集类型 1：番剧 2：电影 3：纪录片 4：国创 5：电视剧 7：综艺
    season_type_name: string; // 剧集类型名称
    selection_style: string; // 分集选择按钮风格，horizontal：横排式 grid：按钮式
    ep_size: number; // 结果匹配的分集数
    url: string; // 剧集重定向url
    button_text: string; // 观看按钮文字
    is_follow: number; // 是否追番
    is_selection: number; // 不知道有什么用途
    eps: {
        id: number; // 分集epid
        cover: string; // 分集封面
        title: string; // 完整标题
        url: string; // 分集重定向url
        release_date: string; // 空
        badges: {
            text: string; // 剧集标志
            text_color: string; // 文字颜色
            text_color_night: string; // 夜间文字颜色
            bg_color: string; // 背景颜色
            bg_color_night: string; // 夜间背景颜色
            border_color: string; // 边框颜色
            border_color_night: string; // 夜间边框颜色
            bg_style: number; // 不知道有什么用途
        }[]; // 分集标志
        index_title: string; // 短标题
        long_title: string; // 单集标题
    }[]; // 结果匹配的分集信息
    cover: string; // 剧集封面
    areas: string; // 地区
    styles: string; // 风格
    goto_url: string; // 剧集重定向url
    desc: string; // 剧集简介
    pubtime: number; // 开播时间
    media_mode: number; // 不知道有什么用途
    fix_pubtime_str: string; // 开播时间重写信息
    media_score: {
        score: number; // 评分
        user_count: number; // 总计评分人数
    }; // 评分

    display_info: {
        text: string;
        text_color: string;
        text_color_night: string;
        bg_color: string;
        bg_color_night: string;
        border_color: string;
        border_color_night: string;
        bg_style: number;
    }[]; // 剧集标志信息
    pgc_season_id: number; // 剧集ssid
    corner: number; // 角标有无	2：无 13：有
    index_show: string; // 全*话
}
// ...
interface web_gameResult
{
    game_base_id: number; // 还未知
    game_name: string; // 游戏名
    game_icon: string; // 游戏图标
    summary: string; // 游戏简介
    game_status: number;
    game_link: string; // 游戏链接
    grade: number;
    book_num: number;
    download_num: number;
    comment_num: number;
    platform: string; // 平台
    notice_title: string; // 公告标题
    notice: string; // 简介?
    game_tags: string; // 游戏标签
    recommend_reason: string; // 不知道有什么用途
    rank_info: {
        rank_type: number; // 不知道有什么用途
        game_rank: number; // 不知道有什么用途
        search_night_icon_url: string; // 不知道有什么用途
        search_day_icon_url: string; // 不知道有什么用途
        search_bkg_night_color: string; // 不知道有什么用途
        search_bkg_day_color: string;  // 不知道有什么用途
        search_font_night_color: string; // 不知道有什么用途
        search_font_day_color: string; // 不知道有什么用途
        rank_content: string; // 不知道有什么用途
        rank_link: string; // 不知道有什么用途
    },
    official_account: number; // 不知道有什么用途
}
// ...
interface liveResult
{
    live_room: live_roomResult[];
    live_user: live_userResult[];
}
// ...
interface cost_time
{
    total: string; // 总耗时
    fetch_lexicon: string; // 不知道有什么用途
    params_check: string; // 不知道有什么用途
    is_risk_query: string; // 不知道有什么用途
    illegal_handler: string; // 不知道有什么用途
    main_handler: string; // 不知道有什么用途
    'get upuser live status'?: string; // 不知道有什么用途
    as_request_format: string; // 不知道有什么用途
    as_request: string; // 不知道有什么用途
    deserialize_response: string; // 不知道有什么用途
    as_response_format: string; // 不知道有什么用途
}

interface exp_list
{
    '5505': boolean;
    '6600': boolean;
    '7707': boolean;
    '9902': boolean;
    '9911': boolean;
    '9924': boolean;
    '9931': boolean;
    '9940': boolean;
    '9961': boolean;
    '9971': boolean;
    '9987': boolean;
    '100206': boolean;
    '100300': boolean;
    '100801': boolean;
    '101405': boolean;
    '101501': boolean;
    '101800': boolean;
    '101901': boolean;
    '102000': boolean;
    '102506': boolean;
    '102600': boolean;
    '102801': boolean;
    '102902': boolean;
    '103000': boolean;
    '103100': boolean;
    '103207': boolean;
    '103301': boolean;
    '103402': boolean;
    '103501': boolean;
}